002 Code/backend/requirements.txt
# FastAPI 및 서버
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 데이터 검증
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Firebase
firebase-admin==6.4.0

# 인증 및 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP 클라이언트 (NAVER OCR API 호출용)
requests==2.31.0
httpx==0.26.0

# 환경 변수 관리
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# 유틸리티
python-dateutil==2.8.2

# 테스트
pytest==8.0.0
pytest-asyncio==0.23.3
